backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-generativeai
beautifulsoup4
requests
feedparser
python-dotenv
pydantic
